add: Pruebas en http para vender productos en una ciudad
- Falta mejorar el formato de la url para no mandar la cantidad ahí sino hacerlo a traves de un formulario de la relacion

+++ Juego-Caravana-Medieval/src/main/java/co/edu/javeriana/caravana_medieval/controller/VenderController.java
@@ -1,7 +1,6 @@
 package co.edu.javeriana.caravana_medieval.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,7 +15,7 @@ public class VenderController {
     private VenderService venderService;
 
     @PutMapping("/productos/{idCaravana}/{idProducto}/{cantidad}")
-    public String venderProducto(@PathVariable Long idCaravana, @PathVariable Long idProducto, @PathVariable Long cantidad) {
+    public String venderProducto(@PathVariable Long idCaravana, @PathVariable Long idProducto, @PathVariable int cantidad) {
         venderService.vender(idCaravana, idProducto, cantidad);
         return "Producto vendido exitosamente";
     }
